+++ R/trx_stats.R
@@ -170,6 +170,8 @@ trx_stats <- function(.data,
     }
   }
 
+  check_split_expose_basis(.data, col_exposure)
+
   start_date <- attr(.data, "start_date")
   end_date <- attr(.data, "end_date")